+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# v0.0.2 八路 2015/07/20
+# v0.0.2 用户模块 2015/07/20
 ## 【增加】
  - index.js 定义ui.router路由配置, 启动逻辑初始化userService.
  - **Sass:** 大部分样式[sass/*] 
@@ -10,7 +10,7 @@
   - userFactory 将源数据解析成用户对象.
   - userService 对获取用户方法封装, 和用户服务初始化.
 
-# v0.0.2 八路 2015/07/20
+# v0.0.2 用户模块 2015/07/20
 ## 【增加】
  - bower.json 跟新至维度调整所需的所有依赖模块.
  - **核心模块:**
@@ -22,10 +22,19 @@
  - **错误模块:**
   - 添加依赖模块ngDialog(对话框指令). 
 
+# v0.0.3 用户模块 2015/07/20
 ## 【修改】
  - **错误模块:**
   - 未登录错误时打开登录界面.
  - **用户模块:**
   - 変更getUser方法为getSessionUser,即获取服务器会话用户.
 
-
+# v0.0.3 用户模块 2015/07/20
+## 【增加】
+ - **核心模块:**
+  - 添加储存初始化维度数据的Cookie的key常量.
+ - **用户模块:**
+  - 为userService增加获取Cookie存储的维度数据getDimeByCookie()方法.
+  - 为userService添加获取指定用户Local数据getLocalRecord(userId)方法.
+  - 为userService添加设置指定用户Local数据setLocalRecord(k, v)方法.
+  - 为User类添加新增记录项addRecordItem(k, v)方法.